Move quantization cpu offload tests out of fastcheck
